Support building fstapi using Microsoft MSVC compiler (#49)
It may be useful to integrate FST waveform dumps into simulators developed with Visual Studio.
It mostly works out-of-the box. The only issue I had to debug was waveforms get corrupted
when size of FST grows larger than 2.1 GB. It turned out off_t is 32bit on Windows.
Replaced it with int64_t when building with MSVC to have consistent behavior with Windows and Linux builds.