Add menu option for saving the current signal list to Tcl script
Sometimes when debugging a complex design with multiple signals it is
useful to have some way of saving the current visible signal list so I
can recover it when opening a new gtkwave instance.

For now, it only supports signals and comments (no blanks or analog
extensions) due to limitations of the Tcl commands available. The
numeric representation and colors are not preserved.