When using Sleep() on windows this API takes milliseconds
The gtkwave function caller is expecting to pass in the number of seconds.

This was causing MSYS2 builds to end up in a tight loop causing UI blocking
and the cursor to be always busy.  This can easily be trigged during a reload
and the underlying file or directory was removed, or the VCD is being
rewritten by another process causing Win32 file locking.  Once triggered the
only reliable exit is to kill the process.

Another consideration is to reevaluate the historical reason why this #ifdef
exists at all, as a quick test compile today inside MSYS2 mingw-w64-x86_64-gcc
shows that the POSIX sleep(unsigned int seconds) is implemented as expected
by the gtkwave caller.

Reference:
https://learn.microsoft.com/en-us/search/?terms=sleep&scope=Desktop
https://learn.microsoft.com/en-us/windows/win32/api/synchapi/nf-synchapi-sleep