Use MapViewOfFileEx for mmap on Windows
This was reported by a user in nickg/nvc#596.  The FST writer emulates
mmap on Windows by copying from the file to a temporary buffer and then
flushing that buffer to disk in fstDestroyMmaps.  The problem here is
that while `xc->curval_mem` is flushed, `xc->valpos_mem` is not.  One
fix would be to duplicate the flushing logic for `valpos_mem` but it
seems better to just use the native Windows MapViewOfFileEx API.  Cygwin
has its own mmap emulation layer so the existing POSIX code should work
fine there.